Separate client data management into contexts
