+++ examples/html-standalone/package.json
@@ -8,10 +8,10 @@
     "preview": "vite preview"
   },
   "dependencies": {
-    "@walletconnect/sign-client": "2.4.4",
+    "@walletconnect/sign-client": "2.4.5",
     "@web3modal/standalone": "2.1.1"
   },
   "devDependencies": {
-    "vite": "4.1.3"
+    "vite": "4.1.4"
   }
 }
