+++ examples/html/package.json
@@ -15,6 +15,6 @@
     "ethers": "5.7.2"
   },
   "devDependencies": {
-    "vite": "4.1.3"
+    "vite": "4.1.4"
   }
 }
