+++ examples/nextjs-standalone/package.json
@@ -6,7 +6,7 @@
     "dev": "next dev"
   },
   "dependencies": {
-    "@walletconnect/sign-client": "2.4.4",
+    "@walletconnect/sign-client": "2.4.5",
     "@web3modal/standalone": "2.1.1",
     "next": "13.1.6",
     "react": "18.2.0",
