+++ package.json
@@ -36,9 +36,9 @@
   ],
   "devDependencies": {
     "@types/node": "18.14.0",
-    "@typescript-eslint/eslint-plugin": "5.52.0",
-    "@typescript-eslint/parser": "5.52.0",
-    "esbuild": "0.17.9",
+    "@typescript-eslint/eslint-plugin": "5.53.0",
+    "@typescript-eslint/parser": "5.53.0",
+    "esbuild": "0.17.10",
     "eslint": "8.34.0",
     "eslint-config-prettier": "8.6.0",
     "lerna": "6.5.1",
