+++ packages/ui/src/contexts/index.ts
@@ -1,0 +1,4 @@
+import './w3m-account-context'
+import './w3m-explorer-context'
+import './w3m-network-context'
+import './w3m-theme-context'
