+++ packages/ui/src/partials/w3m-modal/index.ts
@@ -1,13 +1,4 @@
-import {
-  ClientCtrl,
-  ConfigCtrl,
-  CoreUtil,
-  ExplorerCtrl,
-  ModalCtrl,
-  OptionsCtrl,
-  ThemeCtrl,
-  ToastCtrl
-} from '@web3modal/core'
+import { ModalCtrl, OptionsCtrl } from '@web3modal/core'
 import { html, LitElement } from 'lit'
 import { customElement, state } from 'lit/decorators.js'
 import { classMap } from 'lit/directives/class-map.js'
@@ -24,16 +15,12 @@ export class W3mModal extends LitElement {
 
   // -- state & properties ------------------------------------------- //
   @state() private open = false
-  @state() private preload = true
-  @state() private activeChainId?: number = undefined
 
   // -- lifecycle ---------------------------------------------------- //
   public constructor() {
     super()
-    ThemeUtil.setTheme()
 
-    // Subscribe to modal theme changes
-    this.unsubscribeTheme = ThemeCtrl.subscribe(ThemeUtil.setTheme)
+    // Subscribe to modal state
     this.unsubscribeModal = ModalCtrl.subscribe(modalState => {
       if (modalState.open) {
         this.onOpenModalEvent()
@@ -41,53 +28,14 @@ export class W3mModal extends LitElement {
         this.onCloseModalEvent()
       }
     })
-
-    if (!OptionsCtrl.state.isStandalone) {
-      OptionsCtrl.getAccount()
-      const chain = OptionsCtrl.getSelectedChain()
-      this.activeChainId = chain?.id
-      this.fetchEnsProfile()
-      this.fetchBalance()
-
-      // Subscribe network changes
-      this.unwatchNetwork = ClientCtrl.client().watchNetwork(network => {
-        const newChain = network.chain
-        if (newChain && this.activeChainId !== newChain.id) {
-          OptionsCtrl.setSelectedChain(newChain)
-          this.activeChainId = newChain.id
-          OptionsCtrl.resetBalance()
-          this.fetchBalance()
-        }
-      })
-
-      // Subscribe account changes
-      this.unwatchAccount = ClientCtrl.client().watchAccount(account => {
-        const { address } = OptionsCtrl.state
-        if (account.address !== address) {
-          this.fetchEnsProfile(account.address)
-          this.fetchBalance(account.address)
-        }
-        OptionsCtrl.setAddress(account.address)
-        OptionsCtrl.setIsConnected(account.isConnected)
-      })
-    }
-
-    // Load explorer and image data
-    this.preloadModalData()
   }
 
   public disconnectedCallback() {
     this.unsubscribeModal?.()
-    this.unsubscribeTheme?.()
-    this.unwatchAccount?.()
-    this.unwatchNetwork?.()
   }
 
   // -- private ------------------------------------------------------ //
   private readonly unsubscribeModal?: () => void = undefined
-  private readonly unsubscribeTheme?: () => void = undefined
-  private readonly unwatchAccount?: () => void = undefined
-  private readonly unwatchNetwork?: () => void = undefined
   private abortController?: AbortController = undefined
 
   private get overlayEl() {
@@ -96,50 +44,6 @@ export class W3mModal extends LitElement {
 
   private get containerEl() {
     return UiUtil.getShadowRootElement(this, '.w3m-container')
-  }
-
-  private async fetchEnsProfile(profileAddress?: `0x${string}`) {
-    try {
-      if (ConfigCtrl.state.enableAccountView) {
-        OptionsCtrl.setProfileLoading(true)
-        const address = profileAddress ?? OptionsCtrl.state.address
-        const { id } = ClientCtrl.client().getDefaultChain()
-        if (address && id === 1) {
-          const [name, avatar] = await Promise.all([
-            ClientCtrl.client().fetchEnsName({ address, chainId: 1 }),
-            ClientCtrl.client().fetchEnsAvatar({ address, chainId: 1 })
-          ])
-          if (avatar) {
-            await UiUtil.preloadImage(avatar)
-          }
-          OptionsCtrl.setProfileName(name)
-          OptionsCtrl.setProfileAvatar(avatar)
-        }
-      }
-    } catch (err) {
-      console.error(err)
-      ToastCtrl.openToast(UiUtil.getErrorMessage(err), 'error')
-    } finally {
-      OptionsCtrl.setProfileLoading(false)
-    }
-  }
-
-  private async fetchBalance(balanceAddress?: `0x${string}`) {
-    try {
-      if (ConfigCtrl.state.enableAccountView) {
-        OptionsCtrl.setBalanceLoading(true)
-        const address = balanceAddress ?? OptionsCtrl.state.address
-        if (address) {
-          const balance = await ClientCtrl.client().fetchBalance({ address })
-          OptionsCtrl.setBalance({ amount: balance.formatted, symbol: balance.symbol })
-        }
-      }
-    } catch (err) {
-      console.error(err)
-      ToastCtrl.openToast(UiUtil.getErrorMessage(err), 'error')
-    } finally {
-      OptionsCtrl.setBalanceLoading(false)
-    }
   }
 
   private toggleBodyScroll(enabled: boolean) {
@@ -157,65 +61,6 @@ export class W3mModal extends LitElement {
     }
   }
 
-  private async preloadExplorerData() {
-    const { standaloneChains, chains, walletConnectVersion } = OptionsCtrl.state
-
-    const chainsFilter = standaloneChains?.join(',')
-    await Promise.all([
-      ExplorerCtrl.getPreviewWallets({
-        page: 1,
-        entries: 10,
-        chains: chainsFilter,
-        device: CoreUtil.isMobile() ? 'mobile' : 'desktop',
-        version: walletConnectVersion
-      }),
-      ExplorerCtrl.getRecomendedWallets()
-    ])
-    OptionsCtrl.setIsDataLoaded(true)
-    const { previewWallets, recomendedWallets } = ExplorerCtrl.state
-    const chainsImgs = chains?.map(chain => UiUtil.getChainIcon(chain.id)) ?? []
-    const walletImgs = [...previewWallets, ...recomendedWallets].map(wallet => wallet.image_url.lg)
-    await this.preloadExplorerImages([...chainsImgs, ...walletImgs])
-  }
-
-  private async preloadExplorerImages(images: string[]) {
-    if (images.length) {
-      await Promise.all(images.map(async url => UiUtil.preloadImage(url)))
-    }
-  }
-
-  private async preloadCustomImages() {
-    const images = UiUtil.getCustomImageUrls()
-    if (images.length) {
-      await Promise.all(images.map(async url => UiUtil.preloadImage(url)))
-    }
-  }
-
-  private async preloadConnectorImages() {
-    if (!OptionsCtrl.state.isStandalone) {
-      const images = UiUtil.getConnectorImageUrls()
-      if (images.length) {
-        await Promise.all(images.map(async url => UiUtil.preloadImage(url)))
-      }
-    }
-  }
-
-  private async preloadModalData() {
-    try {
-      if (this.preload) {
-        this.preload = false
-        await Promise.all([
-          this.preloadExplorerData(),
-          this.preloadCustomImages(),
-          this.preloadConnectorImages()
-        ])
-      }
-    } catch (err) {
-      console.error(err)
-      ToastCtrl.openToast('Failed preloading', 'error')
-    }
-  }
-
   private onCloseModal(event: PointerEvent) {
     if (event.target === event.currentTarget) {
       ModalCtrl.close()
@@ -223,7 +68,7 @@ export class W3mModal extends LitElement {
   }
 
   private async onOpenModalEvent() {
-    await this.preloadModalData()
+    // TODO(ilja) await explorer data load here
     this.toggleBodyScroll(false)
     const delay = 0.2
     await animate(this.containerEl, { y: 0 }, { duration: 0 }).finished
@@ -281,6 +126,17 @@ export class W3mModal extends LitElement {
     this.abortController = undefined
   }
 
+  private managedModalContextTemplate() {
+    const { isStandalone } = OptionsCtrl.state
+
+    return isStandalone
+      ? null
+      : html`
+          <w3m-account-context></w3m-account-context>
+          <w3m-network-context></w3m-network-context>
+        `
+  }
+
   // -- render ------------------------------------------------------- //
   protected render() {
     const classes = {
@@ -289,6 +145,11 @@ export class W3mModal extends LitElement {
     }
 
     return html`
+      <w3m-explorer-context></w3m-explorer-context>
+      <w3m-theme-context></w3m-theme-context>
+
+      ${this.managedModalContextTemplate()}
+
       <div
         id="w3m-modal"
         class=${classMap(classes)}
